Added in volatility processing module, and modified the HTML template to show the results of volatility.
If using volatility 2.2 you will need the virtualbox address space extension found here : http://wiki.yobi.be/wiki/RAM_analysis .  Volatility 2.3 includes this by default.  This was not tested on anything other then virtualbox.  Memory dumps must be turned on in cuckoo.conf.  The location of vol.py must be updated with your path in the processing module.  This is really just a quick hack to get volatility to run and as noted in the cuckoo documentation cuckoo may dilute the results.  The api hooking module did not work for me.  Append / delete from the modules_to_run list in volatility.py as needed.